Okay, I've refactored the code as you requested!
Here's what I did:

*   I renamed the `prediction_service.py` file (and its previous name `recommendation_service.py`) to `services/participation_service.py` to better reflect its purpose.
*   I updated all the import statements in your controller files (`dashboard_routes.py`, `program_controller.py`) and test files (`tests/test_participation_service.py`) to point to the new service path.
*   I also renamed the test file to `tests/test_participation_service.py`.
*   Finally, I removed the TODO comment about renaming the file.
